Moved Enums into their own folder
